fetching a dictionary type result with column name
The result we were getting were of list format and difficult to map the column names, added a new method docResult to give a dictionary type result. By default it'll be a list like it is now, but if you give doc=True in fetchall(), fetchmany() or fetchone(), it'll return a dictionary type result.